# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  github_com_stack-service_stack_service_internal_domain_entities.Address:
    properties:
      city:
        type: string
      country:
        type: string
      postalCode:
        type: string
      state:
        type: string
      street:
        type: string
    required:
    - city
    - country
    - postalCode
    - street
    type: object
...
  github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
...
  github_com_stack-service_stack_service_internal_domain_entities.KYCDocument:
    properties:
      contentType:
        type: string
      fileUrl:
        type: string
      type:
        description: passport, drivers_license, etc.
        type: string
    required:
    - contentType
    - fileUrl
    - type
    type: object
  github_com_stack-service_stack_service_internal_domain_entities.KYCPersonalInfo:
    properties:
      address:
        $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.Address'
      country:
        type: string
      dateOfBirth:
        type: string
      firstName:
        type: string
      lastName:
        type: string
    required:
    - country
    - firstName
    - lastName
    type: object
  github_com_stack-service_stack_service_internal_domain_entities.KYCStatusResponse:
    properties:
      approvedAt:
        type: string
      hasSubmitted:
        type: boolean
      lastSubmittedAt:
        type: string
      nextSteps:
        items:
          type: string
        type: array
      providerReference:
        type: string
      rejectionReason:
        type: string
      requiredFor:
        items:
          type: string
        type: array
      requiresKyc:
        type: boolean
      status:
        type: string
      userId:
        type: string
      verified:
        type: boolean
    type: object
  github_com_stack-service_stack_service_internal_domain_entities.KYCSubmitRequest:
    properties:
      documentType:
        type: string
      documents:
        items:
          $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.KYCDocument'
        minItems: 1
        type: array
      metadata:
        additionalProperties: true
        type: object
      personalInfo:
        $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.KYCPersonalInfo'
    required:
    - documentType
    - documents
    type: object
...
  github_com_stack-service_stack_service_internal_domain_entities.OnboardingStartRequest:
    properties:
      email:
        type: string
      phone:
        type: string
    required:
    - email
    type: object
  github_com_stack-service_stack_service_internal_domain_entities.OnboardingStartResponse:
    properties:
      nextStep:
        $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.OnboardingStepType'
      onboardingStatus:
        $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.OnboardingStatus'
      sessionToken:
        type: string
      userId:
        type: string
    type: object
  github_com_stack-service_stack_service_internal_domain_entities.OnboardingStatus:
    enum:
    - started
    - kyc_pending
    - kyc_approved
    - kyc_rejected
    - wallets_pending
    - completed
    type: string
    x-enum-varnames:
    - OnboardingStatusStarted
    - OnboardingStatusKYCPending
    - OnboardingStatusKYCApproved
    - OnboardingStatusKYCRejected
    - OnboardingStatusWalletsPending
    - OnboardingStatusCompleted
  github_com_stack-service_stack_service_internal_domain_entities.OnboardingStatusResponse:
    properties:
      canProceed:
        type: boolean
      completedSteps:
        items:
          $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.OnboardingStepType'
        type: array
      currentStep:
        $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.OnboardingStepType'
      kycStatus:
        type: string
      onboardingStatus:
        $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.OnboardingStatus'
      requiredActions:
        items:
          type: string
        type: array
      userId:
        type: string
      walletStatus:
        $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.WalletStatusSummary'
    type: object
  github_com_stack-service_stack_service_internal_domain_entities.OnboardingStepType:
    enum:
    - registration
    - email_verification
    - phone_verification
    - passcode_creation
    - kyc_submission
    - kyc_review
    - wallet_creation
    - completed
    type: string
    x-enum-varnames:
    - StepRegistration
    - StepEmailVerification
    - StepPhoneVerification
    - StepPasscodeCreation
    - StepKYCSubmission
    - StepKYCReview
    - StepWalletCreation
    - StepOnboardingComplete
...
  github_com_stack-service_stack_service_internal_domain_entities.WalletStatusSummary:
    properties:
      createdWallets:
        type: integer
      failedWallets:
        type: integer
      pendingWallets:
        type: integer
      supportedChains:
        items:
          type: string
        type: array
      totalWallets:
        type: integer
      walletsByChain:
        additionalProperties:
          type: string
        description: chain -> status
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@stackservice.com
    name: API Support
    url: http://www.stackservice.com/support
  description: GenZ Web3 Multi-Chain Investment Platform API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Stack Service API
  version: "1.0"
paths:
...
  /api/v1/kyc/status:
    get:
      description: Returns the user's current KYC verification status and guidance
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.KYCStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get KYC status
      tags:
      - onboarding
  /api/v1/onboarding/kyc/callback/{provider_ref}:
    post:
      consumes:
      - application/json
      description: Handles callbacks from KYC providers with verification results
      parameters:
      - description: KYC provider reference
        in: path
        name: provider_ref
        required: true
        type: string
      - description: KYC callback data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Callback processed
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse'
      summary: Process KYC callback
      tags:
      - onboarding
  /api/v1/onboarding/kyc/submit:
    post:
      consumes:
      - application/json
      description: Submits KYC documents for verification
      parameters:
      - description: KYC submission data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.KYCSubmitRequest'
      produces:
      - application/json
      responses:
        "202":
          description: KYC submission accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse'
        "403":
          description: User not eligible for KYC
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit KYC documents
      tags:
      - onboarding
  /api/v1/onboarding/start:
    post:
      consumes:
      - application/json
      description: Initiates the onboarding process for a new user with email/phone
        verification
      parameters:
      - description: Onboarding start data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.OnboardingStartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.OnboardingStartResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse'
      summary: Start user onboarding
      tags:
      - onboarding
  /api/v1/onboarding/status:
    get:
      description: Returns the current onboarding status for the authenticated user
      parameters:
      - description: User ID (for admin use)
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.OnboardingStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_stack-service_stack_service_internal_domain_entities.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get onboarding status
      tags:
      - onboarding
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"